farmer-chat/requirements.txt
aioboto3==11.2.0
aiobotocore==2.5.0
aiohttp==3.8.4
aioitertools==0.11.0
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
black==24.4.2
boto3==1.26.76
botocore==1.29.76
cachetools==4.2.4
celery==5.3.1
certifi==2023.5.7
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distro==1.9.0
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.2.1
frozenlist==1.4.1
google-api-core==1.34.0
google-api-python-client==2.88.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
google-cloud-core==1.7.3
google-cloud-speech==2.3.0
google-cloud-texttospeech==2.3.0
google-cloud-translate==2.0.2
googleapis-common-protos==1.59.0
grpcio==1.63.0
grpcio-status==1.48.2
grpcio-tools==1.48.2
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
inflection==0.5.1
jmespath==1.0.1
kombu==5.3.7
libcst==1.3.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.24.3
oauthlib==3.2.2
openai==1.3.5
packaging==24.0
pandas==2.0.2
pathspec==0.12.1
peewee==3.16.2
peewee-migrations==0.3.32
Pillow==10.0.0
platformdirs==4.2.1
portalocker==2.8.2
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==3.20.3
psycopg2-binary==2.9.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
qdrant-client==1.8.0
redis==5.0.0
regex==2024.4.28
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.6.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
tensorflow==2.15.0
tomli==2.0.1
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.29.0
vine==5.1.0
wcwidth==0.2.13
wrapt==1.14.0
yarl==1.9.4

